Closes #60
- Changed the PeProvisionerSimple to extend the ResourceProvisionerSimple and removed all the duplicated code.
- Fixed the issue of allocating the same physical PE for multiple Virtual PEs inside the VmSchedulerTimeShared class.
- Changed the signature of the PeSimple constructor that now requires the PE MIPS capacity instead of an ID,
  because it is in fact an attribute that must belong to a PE. The PE that has a MIPS capacity.
- The PeProvisionerSimple constructor doesn't require any parameter at all anymore. It internally creates
   an association with the PE that receives the PeProvisioner instance when a Pe is created.
- All the duplicated code inside the PeProvisioner and PeProvisionerSimple were removed
  and the PeProvisioner is now an interface.
 - The .travis.yml file was updated after the projet was published at Maven Central. By this way, all the modules
   can be tested by travis-ci service (since it download the defined dependencies from maven central, it was not possible
   before that CloudSim Plus was published there to test the modules that depend on it).
   The configuration was update and fixed to always run the tests, including the integration tests.
 - Updated examples.

+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/cloudlets/CloudletExecutionInfo.java
@@ -222,8 +222,8 @@ public class CloudletExecutionInfo {
      * @post $result >= 0
      */
     public long getRemainingCloudletLength() {
-        final long remainingMI = cloudlet.getLength() - (instructionsFinishedSoFar / (long)Conversion.MILLION);
-        return (remainingMI < 0 ? 0 : remainingMI);
+        final double remainingMI = cloudlet.getLength() - (instructionsFinishedSoFar / (double)Conversion.MILLION);
+        return (remainingMI < 0 ? 0 : (long)remainingMI);
     }
 
     /**
